Seperated input and out configs for recon to ease the necessity for dbconnect
