+++ src/databricks/labs/remorph/reconcile/exception.py
@@ -1,5 +1,5 @@
 from pyspark.errors import PySparkException
-from databricks.labs.remorph.reconcile.recon_config import ReconcileOutput
+from databricks.labs.remorph.reconcile.recon_output_conifg import ReconcileOutput
 
 
 class ColumnMismatchException(Exception):
