+++ src/databricks/labs/remorph/reconcile/recon_config.py
@@ -1,10 +1,10 @@
 from __future__ import annotations
 
 import logging
-from collections.abc import Callable
-from dataclasses import dataclass, field
 
-from pyspark.sql import DataFrame
+from dataclasses import dataclass
+from collections.abc import Callable
+
 from sqlglot import expressions as exp
 
 from databricks.labs.remorph.reconcile.constants import SamplingOptionMethod, SamplingSpecificationsType
@@ -31,6 +31,12 @@ class TableThresholdBoundsException(ValueError):
 
 class InvalidModelForTableThreshold(ValueError):
     """Raise the error when the model for table threshold is invalid"""
+
+
+@dataclass
+class HashAlgoMapping:
+    source: Callable
+    target: Callable
 
 
 @dataclass
@@ -284,86 +290,6 @@ class Schema:
 
 
 @dataclass
-class MismatchOutput:
-    mismatch_df: DataFrame | None = None
-    mismatch_columns: list[str] | None = None
-
-
-@dataclass
-class ThresholdOutput:
-    threshold_df: DataFrame | None = None
-    threshold_mismatch_count: int = 0
-
-
-@dataclass
-class DataReconcileOutput:
-    mismatch_count: int = 0
-    missing_in_src_count: int = 0
-    missing_in_tgt_count: int = 0
-    mismatch: MismatchOutput = field(default_factory=MismatchOutput)
-    missing_in_src: DataFrame | None = None
-    missing_in_tgt: DataFrame | None = None
-    threshold_output: ThresholdOutput = field(default_factory=ThresholdOutput)
-    exception: str | None = None
-
-
-@dataclass
-class HashAlgoMapping:
-    source: Callable
-    target: Callable
-
-
-@dataclass
-class SchemaMatchResult:
-    source_column: str
-    source_datatype: str
-    databricks_column: str
-    databricks_datatype: str
-    is_valid: bool = True
-
-
-@dataclass
-class SchemaReconcileOutput:
-    is_valid: bool
-    compare_df: DataFrame | None = None
-    exception: str | None = None
-
-
-@dataclass
-class ReconcileProcessDuration:
-    start_ts: str
-    end_ts: str | None
-
-
-@dataclass
-class StatusOutput:
-    row: bool | None = None
-    column: bool | None = None
-    schema: bool | None = None
-    aggregate: bool | None = None
-
-
-@dataclass
-class ReconcileTableOutput:
-    target_table_name: str
-    source_table_name: str
-    status: StatusOutput = field(default_factory=StatusOutput)
-    exception_message: str | None = None
-
-
-@dataclass
-class ReconcileOutput:
-    recon_id: str
-    results: list[ReconcileTableOutput]
-
-
-@dataclass
-class ReconcileRecordCount:
-    source: int = 0
-    target: int = 0
-
-
-@dataclass
 class Aggregate:
     agg_columns: list[str]
     type: str
@@ -432,9 +358,3 @@ class AggregateQueryRules:
     group_by_columns_as_str: str
     query: str
     rules: list[AggregateRule]
-
-
-@dataclass
-class AggregateQueryOutput:
-    rule: AggregateRule | None
-    reconcile_output: DataReconcileOutput
