+++ src/databricks/labs/remorph/reconcile/schema_compare.py
@@ -6,12 +6,8 @@ from pyspark.sql.types import BooleanType, StringType, StructField, StructType
 from sqlglot import Dialect, parse_one
 
 from databricks.labs.remorph.transpiler.sqlglot.dialect_utils import get_dialect
-from databricks.labs.remorph.reconcile.recon_config import (
-    Schema,
-    SchemaMatchResult,
-    SchemaReconcileOutput,
-    Table,
-)
+from databricks.labs.remorph.reconcile.recon_config import Schema, Table
+from databricks.labs.remorph.reconcile.recon_output_conifg import SchemaMatchResult, SchemaReconcileOutput
 from databricks.labs.remorph.transpiler.sqlglot.generator.databricks import Databricks
 
 logger = logging.getLogger(__name__)
