+++ tests/integration/reconcile/test_recon_capture.py
@@ -15,7 +15,7 @@ from databricks.labs.remorph.reconcile.recon_capture import (
     generate_final_reconcile_output,
     ReconIntermediatePersist,
 )
-from databricks.labs.remorph.reconcile.recon_config import (
+from databricks.labs.remorph.reconcile.recon_output_conifg import (
     DataReconcileOutput,
     MismatchOutput,
     ReconcileOutput,
@@ -23,9 +23,11 @@ from databricks.labs.remorph.reconcile.recon_config import (
     ReconcileTableOutput,
     SchemaReconcileOutput,
     StatusOutput,
-    Table,
     ThresholdOutput,
     ReconcileRecordCount,
+)
+from databricks.labs.remorph.reconcile.recon_config import (
+    Table,
     TableThresholds,
     TableThresholdBoundsException,
 )
